docs: make npm completion stay in sync with npm updates
The current instructions result in a completion script generated by the currently the installed version of npm to be written to the user's shell configuration. If improvements to the script are made in a later npm version, the user will not get these, nor will they know of this condition.

The updated instructions regenerates the script from the currently installed version of npm each time the shell initializes, thereby staying in sync.